RavenDB-23766 - remove the new version available alert only for the server store notification storage
