+++ src/Raven.Server/NotificationCenter/NotificationsStorage.cs
@@ -2,9 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using Raven.Client.Util;
-using Raven.Server.Json;
 using Raven.Server.NotificationCenter.Notifications;
-using Raven.Server.NotificationCenter.Notifications.Details;
 using Raven.Server.Rachis.Commands;
 using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Context;
@@ -360,9 +358,8 @@ namespace Raven.Server.NotificationCenter
             }
         }
 
-        private void Cleanup()
+        protected virtual void Cleanup()
         {
-            RemoveNewVersionAvailableAlertIfNecessary();
         }
 
         private static string GetTableName(string resourceName)
@@ -370,40 +367,6 @@ namespace Raven.Server.NotificationCenter
             return string.IsNullOrEmpty(resourceName)
                 ? Documents.Schemas.Notifications.NotificationsTree
                 : $"{Documents.Schemas.Notifications.NotificationsTree}.{resourceName.ToLowerInvariant()}";
-        }
-
-        private void RemoveNewVersionAvailableAlertIfNecessary()
-        {
-            var buildNumber = ServerVersion.Build;
-
-            var id = AlertRaised.GetKey(AlertType.Server_NewVersionAvailable, null);
-            using (Read(id, out var ntv))
-            {
-                using (ntv)
-                {
-                    if (ntv == null)
-                        return;
-
-                    var delete = true;
-
-                    if (buildNumber != ServerVersion.DevBuildNumber)
-                    {
-                        if (ntv.Json.TryGetMember(nameof(AlertRaised.Details), out var o)
-                            && o is BlittableJsonReaderObject detailsJson)
-                        {
-                            if (detailsJson.TryGetMember(nameof(NewVersionAvailableDetails.VersionInfo), out o)
-                                && o is BlittableJsonReaderObject newVersionDetailsJson)
-                            {
-                                var value = JsonDeserializationServer.LatestVersionCheckVersionInfo(newVersionDetailsJson);
-                                delete = value.BuildNumber <= buildNumber;
-                            }
-                        }
-                    }
-
-                    if (delete)
-                        Delete(id);
-                }
-            }
         }
 
         [DoesNotReturn]
